send only text instead of full html
External content records have a length limit of 10000 characters (and in this action, we conservatively send only 9000 characters). Many of these characters will be HTML tags, which are not useful for searches.

Additionally, we allow setting a CSS selector parameter to narrow down what’s being sent. By default, we are sending the whole `<body>`, but it includes elements like navigation and footer, which makes difficult searching for terms mentioned in them. This value can be set to `main` or `article` to focus on the most relevant part of a page.